requirements.txt
# Flask Backend Requirements  
# Install with: pip install -r requirements.txt

# Build tools (install first)
setuptools>=65.0.0
wheel>=0.38.0

# Web Framework
Flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0

# Machine Learning (compatible with Python 3.12)
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0

# NLP - Indonesian
nltk>=3.8.1
Sastrawi>=1.2.0
regex>=2023.10.3

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0